refactor: move metadb.go to store package
